% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_domain_info}
\alias{get_domain_info}
\title{Get domain info via Datanyze API}
\usage{
get_domain_info(email, token, domain, include_tech = "true")
}
\arguments{
\item{email}{valid email registered in Datanyze}

\item{token}{Datanyze token}

\item{domain}{domain of interest}

\item{include_tech}{defaults to "true", see https://www.datanyze.com/api-documentation/ for more info}
}
\value{
The \code{curl} object with \code{status_code, headers, content} etc.
}
\description{
Get domain info via Datanyze API
}
\examples{
\dontrun{
get_domain_info("my_email", token = "my_token", domain = "facebook.com")
get_domain_info("my_email", token = "my_token", domain = "facebook.com", include_tech = "false")
}
}
