% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{json2formatted_list}
\alias{json2formatted_list}
\title{Generate a formatted list out of Datanyze's response object}
\usage{
json2formatted_list(json_response)
}
\arguments{
\item{json_response}{the \code{curl} object recieved as a response from \code{get_domain_info}}
}
\value{
The \code{list} with selected features - can be coerced to data.frame
}
\description{
Generate a formatted list out of Datanyze's response object
}
\examples{
\dontrun{
json2formatted_list(json_response =
  get_domain_info("my_email", token = "my_token", domain = "facebook.com"))
}
}
